requirements.txt
fastapi==0.110.0
uvicorn==0.27.1
python-multipart==0.0.9
requests==2.31.0
streamlit==1.32.0
pandas==2.1.4
numpy==1.26.4
matplotlib==3.8.2
seaborn==0.13.0
scipy==1.11.4
ydata-profiling==4.10.0
streamlit-pandas-profiling==0.1.3
fpdf2==2.7.9
imbalanced-learn==0.12.0
pycaret==3.3.0
openpyxl==3.1.2
scikit-learn==1.3.2 